changed bootstrap of the app, changed usage of enviroment in favour of environmental variables
