+++ server/.env.example
@@ -1,0 +1,7 @@
+UPLOAD_PATH=uploads
+INTERVAL_TIME=5000
+RETRIES=3
+PAGE_SIZE=15
+DB_URI=mongodb://foouser:foopwd@localhost:27017/kaholo
+PORT=3000
+SERVER_KEY=5843a9f9afd6c9bc4244506ebfb04121
