+++ server/api/services/token.service.js
@@ -1,15 +1,14 @@
-const config = require('../../env/config');
 const jwt = require('jsonwebtoken');
 
 module.exports = {
 
   createToken: (payload, expiration)=> {
-    const jwtSecret = config.serverKey;
+    const jwtSecret = process.env.SERVER_KEY;
     return jwt.sign(payload, jwtSecret, expiration );
   },
 
   validateAndExtractToken: (token)=> {
-    const jwtSecret = config.serverKey;
+    const jwtSecret = process.env.SERVER_KEY;
     try {
       return jwt.verify(token, jwtSecret );
     } catch (err) {
