+++ server/app.js
@@ -1,8 +1,3 @@
-const {app, server, port, winston} = require('./api');
+const startKaholo = require('./helpers');
 
-const bootstrap = require('./helpers/bootstrap').bootstrap;
-
-server.listen(port, () => {
-  winston.log('info', `Running on localhost:${port}`);
-  bootstrap(app);
-});
+startKaholo();
