+++ server/package.json
@@ -49,6 +49,7 @@
     "bluebird": "3.5.1",
     "body-parser": "1.18.2",
     "del": "3.0.0",
+    "dotenv": "^8.2.0",
     "express": "4.16.2",
     "express-winston": "3.0.0",
     "graphlib": "2.1.1",
